+++ app/src/main/java/com/wasingun/seller_lounge/ui/home/HomeFragment.kt
@@ -7,8 +7,7 @@ import androidx.lifecycle.lifecycleScope
 import androidx.navigation.fragment.findNavController
 import com.google.android.material.tabs.TabLayoutMediator
 import com.wasingun.seller_lounge.R
-import com.wasingun.seller_lounge.SellerLoungeApplication
-import com.wasingun.seller_lounge.data.enums.ProductCategory
+import com.wasingun.seller_lounge.data.model.ProductCategory
 import com.wasingun.seller_lounge.databinding.FragmentHomeBinding
 import com.wasingun.seller_lounge.extensions.showTextMessage
 import com.wasingun.seller_lounge.ui.BaseFragment
@@ -51,7 +50,7 @@ class HomeFragment : BaseFragment<FragmentHomeBinding>() {
     }
 
     private fun checkLoginState() {
-        if (SellerLoungeApplication.auth.currentUser == null) {
+        if (sharedViewModel.getCurrentUser() == null) {
             val action = HomeFragmentDirections.actionDestHomeToDestLogin()
             findNavController().navigate(action)
         }
