+++ app/src/main/java/com/wasingun/seller_lounge/ui/home/HomePostFragment.kt
@@ -8,7 +8,7 @@ import androidx.lifecycle.flowWithLifecycle
 import androidx.lifecycle.lifecycleScope
 import androidx.navigation.fragment.findNavController
 import com.wasingun.seller_lounge.R
-import com.wasingun.seller_lounge.data.enums.ProductCategory
+import com.wasingun.seller_lounge.data.model.ProductCategory
 import com.wasingun.seller_lounge.databinding.LayoutHomePostBinding
 import com.wasingun.seller_lounge.extensions.showTextMessage
 import com.wasingun.seller_lounge.ui.BaseFragment
