+++ app/src/main/java/com/wasingun/seller_lounge/ui/home/HomeViewPagerAdapter.kt
@@ -2,7 +2,7 @@ package com.wasingun.seller_lounge.ui.home
 
 import androidx.fragment.app.Fragment
 import androidx.viewpager2.adapter.FragmentStateAdapter
-import com.wasingun.seller_lounge.data.enums.ProductCategory
+import com.wasingun.seller_lounge.data.model.ProductCategory
 
 class HomeViewPagerAdapter(fragment: Fragment) : FragmentStateAdapter(fragment) {
     private val categories = ProductCategory.values()
