+++ app/src/main/java/com/wasingun/seller_lounge/ui/postcontent/DocumentDeleteListener.kt
@@ -1,6 +1,6 @@
 package com.wasingun.seller_lounge.ui.postcontent
 
-import com.wasingun.seller_lounge.data.model.DocumentContent
+import com.wasingun.seller_lounge.data.model.attachedcontent.DocumentContent
 
 fun interface DocumentDeleteListener {
