+++ app/src/main/java/com/wasingun/seller_lounge/ui/postcontent/ImageDeleteListener.kt
@@ -1,6 +1,6 @@
 package com.wasingun.seller_lounge.ui.postcontent
 
-import com.wasingun.seller_lounge.data.model.ImageContent
+import com.wasingun.seller_lounge.data.model.attachedcontent.ImageContent
 
 fun interface ImageDeleteListener {
