+++ app/src/main/java/com/wasingun/seller_lounge/ui/postcontent/PostContentFragment.kt
@@ -12,9 +12,9 @@ import androidx.fragment.app.viewModels
 import androidx.lifecycle.lifecycleScope
 import androidx.navigation.fragment.findNavController
 import com.wasingun.seller_lounge.R
-import com.wasingun.seller_lounge.data.enums.ProductCategory
-import com.wasingun.seller_lounge.data.model.DocumentContent
-import com.wasingun.seller_lounge.data.model.ImageContent
+import com.wasingun.seller_lounge.data.model.ProductCategory
+import com.wasingun.seller_lounge.data.model.attachedcontent.DocumentContent
+import com.wasingun.seller_lounge.data.model.attachedcontent.ImageContent
 import com.wasingun.seller_lounge.databinding.FragmentPostContentBinding
 import com.wasingun.seller_lounge.extensions.showTextMessage
 import com.wasingun.seller_lounge.ui.BaseFragment
